refactor(view): input, output 패키지 생성
